feat: extract file related constants in src/shared/constants
move them to shared/constants/file
