+++ src/app/models/form.server.model.ts
@@ -9,6 +9,7 @@ import mongoose, {
 } from 'mongoose'
 import validator from 'validator'
 
+import { MB } from '../../../shared/constants/file'
 import {
   ADMIN_FORM_META_FIELDS,
   EMAIL_FORM_SETTINGS_FIELDS,
@@ -17,7 +18,6 @@ import {
   STORAGE_PUBLIC_FORM_FIELDS,
 } from '../../../shared/constants/form'
 import { reorder } from '../../../shared/utils/immutable-array-fns'
-import { MB } from '../../shared/constants'
 import { getApplicableIfStates } from '../../shared/util/logic'
 import {
   AuthType,
