+++ src/app/modules/submission/email-submission/email-submission.receiver.ts
@@ -2,7 +2,7 @@ import Busboy from 'busboy'
 import { IncomingHttpHeaders } from 'http'
 import { err, ok, Result, ResultAsync } from 'neverthrow'
 
-import { MB } from '../../../../shared/constants'
+import { MB } from '../../../../../shared/constants/file'
 import { IAttachmentInfo } from '../../../../types'
 import { createLoggerWithLabel } from '../../../config/logger'
