+++ src/app/routes/api/v3/admin/forms/__tests__/admin-forms.presign.routes.spec.ts
@@ -7,7 +7,6 @@ import supertest, { Session } from 'supertest-session'
 import { aws } from 'src/app/config/config'
 import { getEncryptedFormModel } from 'src/app/models/form.server.model'
 import getUserModel from 'src/app/models/user.server.model'
-import { VALID_UPLOAD_FILE_TYPES } from 'src/shared/constants'
 import { IUserSchema, ResponseMode, Status } from 'src/types'
 
 import { createAuthedSession } from 'tests/integration/helpers/express-auth'
@@ -15,6 +14,7 @@ import { setupApp } from 'tests/integration/helpers/express-setup'
 import { buildCelebrateError } from 'tests/unit/backend/helpers/celebrate'
 import dbHandler from 'tests/unit/backend/helpers/jest-db'
 
+import { VALID_UPLOAD_FILE_TYPES } from '../../../../../../../../shared/constants/file'
 import { AdminFormsRouter } from '../admin-forms.routes'
 
 // Prevent rate limiting.
